set ehcache version to 2.11.0

+++ ehcache-core/src/main/java/net/sf/ehcache/distribution/ConfigurableRMIClientSocketFactory.java
@@ -1,17 +1,17 @@
 /**
- *  Copyright Terracotta, Inc.
+ * Copyright Terracotta, Inc.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
+ * <a href="http://www.apache.org/licenses/LICENSE-2.0">http://www.apache.org/licenses/LICENSE-2.0</a>
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 package net.sf.ehcache.distribution;
@@ -43,14 +43,13 @@ public final class ConfigurableRMIClientSocketFactory implements Serializable, R
     /**
      * Construct a new socket factory with the given timeout.
      *
-     * @param socketTimeoutMillis
      * @see Socket#setSoTimeout
      */
     public ConfigurableRMIClientSocketFactory(Integer socketTimeoutMillis) {
         if (socketTimeoutMillis == null) {
             this.socketTimeoutMillis = ONE_SECOND;
         } else {
-            this.socketTimeoutMillis = socketTimeoutMillis.intValue();
+            this.socketTimeoutMillis = socketTimeoutMillis;
         }
     }
 
@@ -96,7 +95,7 @@ public final class ConfigurableRMIClientSocketFactory implements Serializable, R
             return false;
         } else {
             return (getClass() == object.getClass() &&
-                socketTimeoutMillis == ((ConfigurableRMIClientSocketFactory) object).socketTimeoutMillis);
+                    socketTimeoutMillis == ((ConfigurableRMIClientSocketFactory) object).socketTimeoutMillis);
         }
     }
